% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_ticket.r
\name{auth_ticket}
\alias{auth_ticket}
\title{A Function for getting Quickbase Authentication Tickets}
\usage{
auth_ticket(ticket_handle)
}
\arguments{
\item{ticket_handle}{A handle object returned by the auth_ticket_handle() function}
}
\description{
This Function uses GET to return a response from quickbase with the Quickbase Ticket ID located within the contents of the html response
}
\keyword{auth_ticket}

