% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_ticket.r
\name{auth_ticket_handle}
\alias{auth_ticket_handle}
\title{A Function for creating URL handle object to retrieve a quickbase ticket id}
\usage{
auth_ticket_handle(URL, username, password, hours = 24)
}
\arguments{
\item{URL}{A url for Quickbase company domain}

\item{username}{Quickbase username's email address}

\item{password}{Username's password}

\item{hours}{Length of time authentication ticket remains valid. Defaults to 24}
}
\description{
This Function passes a character string to the httr::handle function to return a list
}
\keyword{auth_ticket_handle}

