% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ids.R
\name{db_ids}
\alias{db_ids}
\title{GET Quickbase db/app/table names and unique IDs}
\usage{
db_ids(URL, qb_ticket_id)
}
\arguments{
\item{URL}{A url for Quickbase company domain}

\item{qb_ticket_id}{A valid quickbase authentication ticket}
}
\description{
This function sends a request to quickbase to return a response with the db names and ids
which the user has access to. The response in returned as html which is then parsed and returned
as a dataframe.
}
\keyword{db_ids}

