% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_do_query.r
\name{qb_do_query}
\alias{qb_do_query}
\title{GET Request for tablular data within quickbase app.}
\usage{
qb_do_query(URL, db_id, qb_ticket_id, app_token)
}
\arguments{
\item{URL}{A url for Quickbase company domain}

\item{qb_ticket_id}{A valid quickbase authentication ticket}

\item{app_token}{A valid quickbase app token}
}
\description{
This function sends a request to quickbase to return a response with the app data. The response in returned as html which is then parsed and returned
as a html page which then can be parsed using qb_parse
}
\keyword{db_do_query}

