% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_do_query.r
\name{qb_parse}
\alias{qb_parse}
\title{Parses a valid quickbase html do_query response page into a dataframe}
\usage{
qb_parse(html_qb_db)
}
\arguments{
\item{html_qb_db}{A valid quickbase response page return from qb_do_query}
}
\description{
This function takes a page returned by qb_do_query and parses the contents to an R dataframe
}
\keyword{qb_parse}

